feat(dayjs): use custom hook to init dayjs
